requirements.txt
streamlit
langchain
langchain-openai
langgraph==0.3.1
openai
faiss-cpu
pypdf
python-dotenv
pandas
numpy>=1.24.0
tiktoken
requests
psutil
pydantic
langchain_community